implemented the comments
