+++ src/types/Errors.ts
@@ -1,0 +1,8 @@
+export enum Error {
+  NO_ERROR = '',
+  LOAD = 'Unable to load todos',
+  TITLE = 'Title should not be empty',
+  ADD = 'Unable to add a todo',
+  DELETE = 'Unable to delete a todo',
+  UPDATE = 'Unable to update a todo',
+}
