+++ src/types/Filter.ts
@@ -1,0 +1,5 @@
+export enum FilterBy {
+  ALL = 'all',
+  ACTIVE = 'active',
+  COMPLETED = 'comleted',
+}
