+++ src/types/Todo.ts
@@ -4,18 +4,3 @@ export interface Todo {
   title: string;
   completed: boolean;
 }
-
-export enum Error {
-  DEFAULT = '',
-  LOAD = 'Unable to load todos',
-  TITLE = 'Title should not be empty',
-  ADD = 'Unable to add a todo',
-  DELETE = 'Unable to delete a todo',
-  UPDATE = 'Unable to update a todo',
-}
-
-export enum FilterBy {
-  ALL = 'all',
-  ACTIVE = 'active',
-  COMPLETED = 'comleted',
-}
