+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export type { Todo } from './Todo';
+export { Error } from './Errors';
+export { FilterBy } from './Filter';
